+++ package.json
@@ -53,13 +53,13 @@
     "express-validator": "^6.1.1",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
+    "mysql2": "^1.6.5",
     "nodemailer": "^6.3.0",
     "nodemon": "^1.18.11",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "pg": "^7.12.1",
     "pg-hstore": "^2.3.3",
-    "mysql2": "^1.6.5",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-redux": "7.0.3",
@@ -72,14 +72,12 @@
     "sequelize-cli": "^5.5.1"
   },
   "devDependencies": {
-    "babel-jest": "^24.9.0",
-    "identity-obj-proxy": "^3.0.0",
-    "react-test-renderer": "^16.9.0",
     "@babel/core": "^7.0.0",
     "@babel/plugin-proposal-class-properties": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^10.0.2",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.0",
     "clean-webpack-plugin": "^1.0.0",
     "css-loader": "^2.0.0",
@@ -95,6 +93,7 @@
     "eslint-plugin-react": "^7.14.2",
     "eslint-plugin-standard": "^4.0.0",
     "html-webpack-plugin": "^3.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^24.9.0",
     "jest-mock-axios": "^3.1.1",
     "less": "^3.9.0",
@@ -102,10 +101,13 @@
     "moxios": "^0.4.0",
     "npm-run-all": "^4.1.5",
     "null-loader": "^3.0.0",
+    "react-test-renderer": "^16.9.0",
     "redux-devtools-extension": "^2.13.8",
     "redux-mock-store": "^1.5.3",
     "sinon": "^7.4.2",
     "style-loader": "^0.23.0",
+    "superagent": "^5.1.0",
+    "supertest": "^4.0.2",
     "url-loader": "^1.0.1",
     "webpack": "^4.33.0",
     "webpack-cli": "^3.3.7",
