+++ src/server/__tests__/app.test.js
@@ -1,0 +1,11 @@
+const request = require('supertest')
+const app = require('../app.js')
+
+describe('Test the root path', () => {
+  test('It should response the GET method', (done) => {
+    request(app).get('/').then((response) => {
+      expect(response.statusCode).toBe(200)
+      done()
+    })
+  })
+})
